Move `getDependencies()` from S4328 to helpers
